scripts: Deploy scripts and generate calldata
